software/bus-debugger/requirements.txt
paho-mqtt
rich